# app/model/entity/todo.gen.go
// Code generated by gorm.io/gen. DO NOT EDIT.
// Code generated by gorm.io/gen. DO NOT EDIT.
// Code generated by gorm.io/gen. DO NOT EDIT.

package entity

import (
	"time"
)

const TableNameTodo = "todo"

// Todo mapped from table <todo>
type Todo struct {
	ID        int        `gorm:"column:id;type:int;primaryKey;autoIncrement:true" json:"id"`
	UID       int        `gorm:"column:uid;type:int;not null" json:"uid"`
	Priority  int        `gorm:"column:priority;type:tinyint;not null" json:"priority"`                                // 优先级
	Done      int        `gorm:"column:done;type:tinyint;not null" json:"done"`                                        // 完成情况
	Title     string     `gorm:"column:title;type:varchar(255);not null" json:"title"`                                 // 标题
	Content   *string    `gorm:"column:content;type:varchar(255)" json:"content"`                                      // 内容
	StartTime *time.Time `gorm:"column:start_time;type:datetime" json:"start_time" example:"2022-11-16T16:00:00.000Z"` // 日期
}

// TableName Todo's table name
func (*Todo) TableName() string {
	return TableNameTodo
}